Enhance/requirement.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
assemblyai==0.15.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
elevenlabs==0.2.21
executing==1.2.0
fastapi==0.100.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gradio==3.39.0
gradio_client==0.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
llvmlite==0.40.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==10.0.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230314
orjson==3.9.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.5.0
rpds-py==0.9.2
safetensors==0.3.1
scipy==1.11.1
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tdqm==0.0.1
tiktoken==0.3.1
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.1
wcwidth==0.2.6
websockets==11.0.3
whisper-mic==1.1.1
yarl==1.9.2
zipp==3.16.2
